refactor(grids): extract validating constants into const module
